Forcing Unit Tests to choose same first node as production
This is a problem because the first choice is arbitrary and production
does not order the list it chooses from.  By making our first arbitrary force
to use the same as production, then all the params from production match
our unit tests!

I also cleaned up the debug print statements.